update logic into folders service

+++ src/services/folder.service.ts
@@ -1,4 +1,6 @@
+import { PrismaClientKnownRequestError } from "@prisma/client/runtime/library";
 import prisma from "../db";
+import { AppError } from "../types/appError";
 import { FolderCreateBody } from "../types/folder";
 
 export const getAll = async (userId: string) => {
@@ -19,16 +21,22 @@ export const getById = async (userId: string, id: string) => {
 };
 
 export const add = async (userId: string, data: FolderCreateBody) => {
-  const count = await prisma.folder.count({
-    where: {
-      name: data.name,
-    },
-  });
-  if (count > 0) throw new Error("Folder already exists");
-
-  return await prisma.folder.create({
-    data: { ...data, userId },
-  });
+  try {
+    return await prisma.folder.create({
+      data: { ...data, userId },
+    });
+  } catch (error) {
+    if (
+      error instanceof PrismaClientKnownRequestError &&
+      error.code === "P2002"
+    ) {
+      throw new AppError(
+        "A folder with the same name already exists for this user.",
+        409
+      );
+    }
+    throw error;
+  }
 };
 
 export const update = async (
@@ -36,15 +44,47 @@ export const update = async (
   id: string,
   data: FolderCreateBody
 ) => {
-  return await prisma.folder.update({
-    where: {
-      id,
-    },
-    data: { ...data, userId },
-  });
+  let folder = await getById(userId, id);
+
+  if (!folder) {
+    throw new AppError("Folder not found", 404);
+  }
+
+  try {
+    folder = await prisma.folder.update({
+      where: {
+        id,
+      },
+      data: { ...data, userId },
+    });
+  } catch (error) {
+    if (
+      error instanceof PrismaClientKnownRequestError &&
+      error.code === "P2002"
+    ) {
+      throw new AppError(
+        "A folder with the same name already exists for this user.",
+        409
+      );
+    }
+    throw error;
+  }
+  return folder;
 };
 
 export const deleteById = async (userId: string, id: string) => {
+  const defaultFolders = ["Favorites", "Sent", "Drafts", "Trash"];
+
+  const folder = await getById(userId, id);
+
+  if (!folder) {
+    throw new AppError("Folder not found", 404);
+  }
+
+  if (defaultFolders.includes(folder.name)) {
+    throw new AppError("Default folders cannot be deleted", 403);
+  }
+
   if (await getById(userId, id)) {
     await prisma.folder.delete({
       where: {
